move to new Util, alien sprite rotation fix
